[#3] 불필요한 클래스 제거
