--- src/main/java/com/atoz/authentication/AuthenticationConst.java
@@ -1,5 +1,0 @@
-package com.atoz.authentication;
-
-public interface AuthenticationConst {
-    String SIGNIN_MEMBER = "signinMember";
-}
