--- src/main/java/com/atoz/cryptography/HashManager.java
@@ -1,27 +1,0 @@
-package com.atoz.cryptography;
-
-import java.nio.charset.StandardCharsets;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
-
-public class HashManager {
-
-    public byte[] makeSalt() {
-        byte[] salt = new byte[32];
-        SecureRandom random = new SecureRandom();
-        random.nextBytes(salt);
-        return salt;
-    }
-
-    public byte[] computeHash(String input, byte[] salt) {
-        MessageDigest messageDigest;
-        try {
-            messageDigest = MessageDigest.getInstance("SHA-256");
-            messageDigest.update(salt);
-            return messageDigest.digest(input.getBytes(StandardCharsets.UTF_8));
-        } catch (NoSuchAlgorithmException e) {
-            throw new RuntimeException(e);
-        }
-    }
-}
